feat: add install template, needs more features
