--- renovate.json
@@ -1,17 +1,0 @@
-{
-	"$schema": "https://docs.renovatebot.com/renovate-schema.json",
-	"extends": ["config:base"],
-	"major": {
-		"dependencyDashboardApproval": true
-	},
-	"schedule": ["every weekend"],
-	"lockFileMaintenance": {
-		"enabled": true
-	},
-	"packageRules": [
-		{
-			"matchUpdateTypes": ["minor", "patch"],
-			"matchCurrentVersion": "!/^0/"
-		}
-	]
-}
